+++ packages/docs/src/stories/Text.stories.tsx
@@ -1,5 +1,5 @@
 import type { StoryObj, Meta } from '@storybook/react'
-import { Text, TextProps } from '@codingmage-ignite-ui/react'
+import { Text, TextProps } from '@codingmage-ignite-ds/react'
 
 export default {
   title: 'Typography/Text',