+++ packages/eslint-config/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @codingmage-ignite-ui/eslint-config
+# @codingmage-ignite-ds/eslint-config
 
 ## 1.1.0
 
 ### Minor Changes
 
-- 654b5bf: Change from @ignite-ui to @codingmage-ignite-ui
+- 654b5bf: Change from @ignite-ui to @codingmage-ignite-ds