+++ packages/react/CHANGELOG.md
@@ -1,7 +1,13 @@
-# @codingmage-ignite-ui/react
+# @codingmage-ignite-ds/react
+
+## 2.1.0
+
+### Minor Changes
+
+- Add test
 
 ## 2.0.0
 
 ### Major Changes
 
-- 654b5bf: Change from @ignite-ui to @codingmage-ignite-ui
+- 654b5bf: Change from @ignite-ui to @codingmage-ignite-ds