+++ packages/tokens/CHANGELOG.md
@@ -1,10 +1,22 @@
-# @codingmage-ignite-ui/tokens
+# @codingmage-ignite-ds/tokens
+
+## 2.2.0
+
+### Minor Changes
+
+- Add test
+
+## 2.1.0
+
+### Minor Changes
+
+- Add blue test color
 
 ## 2.0.0
 
 ### Major Changes
 
-- 654b5bf: Change from @ignite-ui to @codingmage-ignite-ui
+- 654b5bf: Change from @ignite-ui to @codingmage-ignite-ds
 
 ### Minor Changes